deploy/requirements-rk3328.txt
cryptography==3.4.8
flask==2.0.3
click==8.0.4
requests==2.27.1
pyopenssl==21.0.0
pyyaml==6.0
python-dotenv==0.20.0
hidapi==0.14.0
werkzeug==2.0.3
jinja2==3.0.3
markupsafe==2.0.1
itsdangerous==2.0.1
cffi==1.15.1
pycparser==2.21
urllib3==1.26.18
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.4
